Rename base voltage color config module
