+++ single-line-diagram-color/src/main/java/com/powsybl/sld/color/BaseVoltageColor.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.basevoltage;
+package com.powsybl.sld.color;
 
 import java.io.IOException;
 import java.io.InputStream;
