+++ single-line-diagram-color/src/main/java/com/powsybl/sld/color/BaseVoltageConfig.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.basevoltage;
+package com.powsybl.sld.color;
 
 /**
 *
