+++ single-line-diagram-color/src/test/resources/base-voltages.yml
